#25 Created wallet module and moved existing wallet functionalities.
