+++ settings.gradle
@@ -5,3 +5,4 @@ include 'core'
 include 'crypto'
 include 'model'
 include 'utils'
+include 'wallet'
