+++ wallet/src/test/java/io/neow3j/wallet/WalletUtilsTest.java
@@ -1,5 +1,9 @@
-package io.neow3j.crypto;
+package io.neow3j.wallet;
 
+import io.neow3j.crypto.Credentials;
+import io.neow3j.crypto.ECKeyPair;
+import io.neow3j.crypto.Hash;
+import io.neow3j.crypto.MnemonicUtils;
 import io.neow3j.utils.Numeric;
 import org.junit.After;
 import org.junit.Before;
